# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Generated API",
    "description": "REST API generated by Backend Creator (bcm)",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/v1/bookingRequests": {
      "get": {
        "tags": [
          "BookingRequest"
        ],
        "summary": "List all bookingRequests",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Search term for @bcm.searchable fields"
          },
          {
            "name": "include",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated relation names to include"
          }
        ],
        "responses": {
          "200": {
            "description": "List of bookingRequests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/BookingRequestResponse"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "BookingRequest"
        ],
        "summary": "Create a new bookingRequest",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingRequestCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "BookingRequest created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingRequestDataResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — requires role: \"USER\", \"ADMIN\""
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/bookingRequests/{id}": {
      "get": {
        "tags": [
          "BookingRequest"
        ],
        "summary": "Get bookingRequest by key",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "BookingRequest details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingRequestDataResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "tags": [
          "BookingRequest"
        ],
        "summary": "Update bookingRequest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingRequestUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "BookingRequest updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingRequestDataResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — requires role: \"USER\", \"ADMIN\""
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "BookingRequest"
        ],
        "summary": "Partially update bookingRequest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingRequestPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "BookingRequest updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingRequestDataResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — requires role: \"USER\", \"ADMIN\""
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "BookingRequest"
        ],
        "summary": "Delete bookingRequest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — requires role: \"USER\", \"ADMIN\""
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BookingRequest_UnitInput": {
        "type": "object",
        "properties": {
          "create": {
            "type": "object",
            "properties": {
              "title": {
                "type": "string"
              },
              "description": {
                "type": "string"
              },
              "keywords": {
                "type": "string",
                "nullable": true
              },
              "price": {
                "type": "number",
                "format": "double"
              },
              "availability": {
                "$ref": "#/components/schemas/UnitAvailability"
              },
              "type": {
                "$ref": "#/components/schemas/UnitType"
              },
              "imageUrls": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "videoUrls": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "locationId": {
                "type": "string"
              },
              "acceptedSellRequestId": {
                "type": "string",
                "nullable": true
              },
              "deletedAt": {
                "type": "string",
                "format": "date-time",
                "nullable": true
              }
            },
            "required": [
              "title",
              "description",
              "price",
              "type",
              "locationId"
            ],
            "description": "Create a related record"
          },
          "connect": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              }
            },
            "required": [
              "id"
            ],
            "description": "Connect to an existing record by unique selector"
          }
        },
        "description": "Nested input: provide either 'create' or 'connect'",
        "anyOf": [
          {
            "required": [
              "create"
            ]
          },
          {
            "required": [
              "connect"
            ]
          }
        ]
      },
      "BookingRequest_UserInput": {
        "type": "object",
        "properties": {
          "create": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "password": {
                "type": "string"
              },
              "phone": {
                "type": "string",
                "nullable": true
              },
              "address": {
                "type": "string",
                "nullable": true
              },
              "rate": {
                "type": "number",
                "format": "float"
              },
              "role": {
                "$ref": "#/components/schemas/Role"
              },
              "isVerified": {
                "type": "boolean"
              }
            },
            "required": [
              "name",
              "email",
              "password"
            ],
            "description": "Create a related record"
          },
          "connect": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              }
            },
            "required": [
              "id"
            ],
            "description": "Connect to an existing record by unique selector"
          }
        },
        "description": "Nested input: provide either 'create' or 'connect'",
        "anyOf": [
          {
            "required": [
              "create"
            ]
          },
          {
            "required": [
              "connect"
            ]
          }
        ]
      },
      "BookingRequestResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "unitId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "details": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/RequestStatus"
          },
          "adminNote": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BookingRequestDataResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/BookingRequestResponse"
          }
        }
      },
      "BookingRequestCreate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "details": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/RequestStatus"
          },
          "adminNote": {
            "type": "string",
            "nullable": true
          },
          "unit": {
            "$ref": "#/components/schemas/BookingRequest_UnitInput"
          },
          "user": {
            "$ref": "#/components/schemas/BookingRequest_UserInput"
          }
        },
        "required": [
          "unit",
          "user"
        ]
      },
      "BookingRequestUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "details": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/RequestStatus"
          },
          "adminNote": {
            "type": "string",
            "nullable": true
          },
          "unit": {
            "$ref": "#/components/schemas/BookingRequest_UnitInput"
          },
          "user": {
            "$ref": "#/components/schemas/BookingRequest_UserInput"
          }
        },
        "required": [
          "unit",
          "user"
        ]
      },
      "BookingRequestPatch": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "details": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/RequestStatus"
          },
          "adminNote": {
            "type": "string",
            "nullable": true
          },
          "unit": {
            "$ref": "#/components/schemas/BookingRequest_UnitInput"
          },
          "user": {
            "$ref": "#/components/schemas/BookingRequest_UserInput"
          }
        }
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "USER",
          "ADMIN"
        ]
      },
      "RequestStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "ACCEPTED",
          "REJECTED",
          "CANCELLED",
          "AVAILABLE"
        ]
      },
      "UnitType": {
        "type": "string",
        "enum": [
          "RESIDENTIAL",
          "COMMERCIAL"
        ]
      },
      "UnitAvailability": {
        "type": "string",
        "enum": [
          "AVAILABLE",
          "RESERVED",
          "SOLD",
          "UNAVAILABLE"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}